FileInfoWebApp/src/main/java/com/nuance/file/info/beans/FileOrDirInformation.java
package com.nuance.file.info.beans;

public class FileOrDirInformation {
	
	private String fullPath;
	private String fileSize;
	private String fileName;
	private boolean isDir;

	public String getFullPath() {
		return fullPath;
	}

	public void setFullPath(String fullPath) {
		this.fullPath = fullPath;
	}

	public String getFileSize() {
		return fileSize;
	}

	public void setFileSize(String fileSize) {
		this.fileSize = fileSize;
	}

	public boolean isDir() {
		return isDir;
	}

	public void setDir(boolean isDir) {
		this.isDir = isDir;
	}

	public String getFileName() {
		return fileName;
	}

	public void setFileName(String fileName) {
		this.fileName = fileName;
	}
	
}
